Fix make clean rule (#50)
* Fix make clean rule

The make clean rule was missing a "then", so it wouldn't run properly.
This commit adds it, so now it works as it should.

* Send attachments in the order they were specified

Send attachments in the same order as they are specified in the command
line. They were previously sent in reverse order.

Fixes #13

* Send vcard as the last attachment

If any other attachments are attached, send the vcard as the last one.

* Remove mention to SIGNATURE_DIVIDE from manpage

SIGNATURE_DIVIDE is an obsolete option.